+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.1.0-alpha+2
+
+* Added support for `channels.createMessage`.
+* Added support for `users.getCurrentUser`.
+* Fixed a bug where JSON payloads did not have the right `Content-Type`.
+* Added the `User` and `Message` structures.
+
 ## 0.1.0-alpha+1
 
 * Small change to `REAMDE.md`.